migrate to new adapter api
