+++ app/src/main/java/com/semisonfire/cloudgallery/ui/disk/DiskResult.kt
@@ -1,14 +1,15 @@
 package com.semisonfire.cloudgallery.ui.disk
 
+import com.semisonfire.cloudgallery.adapter.holder.Item
 import com.semisonfire.cloudgallery.core.data.model.Photo
 
 sealed class DiskResult {
 
-    data class Loaded(val photos: List<Photo>) : DiskResult()
+    data class Loaded(val photos: List<Item>) : DiskResult()
     data class Uploading(val photos: List<Photo>) : DiskResult()
     data class PhotoUploaded(val photo: Photo, val uploaded: Boolean) : DiskResult()
     data class PhotoDownloaded(val path: String) : DiskResult()
     data class PhotoDeleted(val photo: Photo) : DiskResult()
 
-    data class LoadMoreCompleted(val photos: List<Photo>) : DiskResult()
+    data class LoadMoreCompleted(val photos: List<Item>) : DiskResult()
 }
