+++ app/src/main/java/com/semisonfire/cloudgallery/ui/disk/data/DiskRepository.kt
@@ -1,5 +1,6 @@
 package com.semisonfire.cloudgallery.ui.disk.data
 
+import com.semisonfire.cloudgallery.adapter.holder.Item
 import com.semisonfire.cloudgallery.core.data.model.Photo
 import com.semisonfire.cloudgallery.core.data.remote.api.DiskApi
 import com.semisonfire.cloudgallery.core.data.remote.api.IMAGE_MEDIA_TYPE
@@ -8,17 +9,18 @@ import com.semisonfire.cloudgallery.core.data.remote.api.SORT_MODIFIED_DESC
 import com.semisonfire.cloudgallery.ui.disk.model.remote.Link
 import io.reactivex.Observable
 import io.reactivex.Single
+import java.io.File
+import javax.inject.Inject
 import okhttp3.MediaType
 import okhttp3.MultipartBody
 import okhttp3.RequestBody
-import java.io.File
-import javax.inject.Inject
 
 class DiskRepository @Inject constructor(
-    private val diskApi: DiskApi
+    private val diskApi: DiskApi,
+    private val mapper: DiskMapper
 ) {
 
-    fun getPhotos(page: Int, limit: Int): Single<List<Photo>> {
+    fun getPhotos(page: Int, limit: Int): Single<List<Item>> {
         return diskApi
             .getDiskImages(
                 limit,
@@ -27,7 +29,9 @@ class DiskRepository @Inject constructor(
                 IMAGE_SIZE_XL,
                 SORT_MODIFIED_DESC
             )
-            .map { it.items ?: emptyList() }
+            .map { response ->
+                response.items?.let { mapper.map(it) } ?: emptyList()
+            }
             .onErrorReturn { emptyList() }
     }
